% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_steps.R
\name{calibration_step1}
\alias{calibration_step1}
\title{First calibration step}
\usage{
calibration_step1(calib_object)
}
\arguments{
\item{calib_object}{a formatted calibration object}
}
\description{
First calibration step
}
