% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_steps.R
\name{calibration_step3}
\alias{calibration_step3}
\title{Third calibration step: Wrap up the calibration system and store the results}
\usage{
calibration_step3(calib_object)
}
\arguments{
\item{calib_object}{a formatted calibration object}
}
\description{
Having this as a separate step allows a mail to be send at the end of the
calibration
}
