% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sideload.R
\name{load_sideload}
\alias{load_sideload}
\title{Read some data saved to be reused by the calibration process}
\usage{
load_sideload(calib_object, job)
}
\description{
Read some data saved to be reused by the calibration process
}
