% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposers.R
\name{make_shrink_proposer}
\alias{make_shrink_proposer}
\title{Shrink the range by a shrink factor (default 2)}
\usage{
make_shrink_proposer(n_new, shrink = 2)
}
\description{
If centers are provided in a sideload, use that
}
