% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment_rmd.R
\name{render_assessment}
\alias{render_assessment}
\title{Generate an html report of the auto-calibration}
\usage{
render_assessment(
  path_to_assessments,
  output_filename = "assessment.html",
  output_dir = NULL
)
}
\arguments{
\item{path_to_assessments}{Path to an \code{assessments.rds} file generated by an
\code{swfcalib} process.}

\item{output_filename}{Name of the html report (default = "assessment.html")}

\item{output_dir}{Directory where to store the report (default = current
working directory)}
}
\description{
The report contains descriptions of the parameters spaces and residual errors
over the duration of the calibration.
}
