% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sideload.R
\name{save_sideload}
\alias{save_sideload}
\title{Save some data to be reused by the calibration process}
\usage{
save_sideload(calib_object, job, x)
}
\description{
Save some data to be reused by the calibration process
}
