% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_sideload}
\alias{save_sideload}
\title{Save some data to be reused by the calibration process}
\usage{
save_sideload(calib_object, x, id)
}
\description{
Save some data to be reused by the calibration process
}
